refactor: centralise TTL constants
